@@ -6,17 +6,17 @@
     "@babel/core": "^7.14.0",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "@reduxjs/toolkit": "^1.5.1",
+    "@reduxjs/toolkit": "^1.9.1",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.2",
     "depcheck": "^1.3.1",
     "moment": "^2.29.1",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-day-picker": "^7.4.10",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.1.0",
-    "react-redux": "^7.2.9",
+    "react-redux": "^8.0.5",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1"
   },
